`abs_diff_eq` broken into sub modules for ease of navigation and separation of concerns
