+++ approx/src/abs_diff_eq/mod.rs
@@ -1,0 +1,9 @@
+mod collections;
+mod core_impls;
+mod external;
+mod primitives;
+mod trait_def;
+#[cfg(feature = "tuple_impl")]
+mod tuples;
+
+pub use trait_def::AbsDiffEq;
